requirements.txt
requests
PyYAML
python-dateutil
tqdm
Jinja2
feedparser
